Edited iOS implementation

+++ google_places_ios/ios/Classes/PlacePhotoExtensions.swift
@@ -3,7 +3,8 @@ import GooglePlaces
 extension GMSPlacePhotoMetadata{
     func toJson() -> Dictionary<String, Any?>{
         return ["attributions": attributions?.string,
-                "width": maxSize.width,
-                "height": maxSize.height]
+                "width": UInt(maxSize.width),
+                "height": UInt(maxSize.height)
+                ]
     }
 }
